Move buildApiEndpoint
