+++ modules/carto/src/sources/utils.ts
@@ -1,7 +1,6 @@
 import {CartoAPIError} from '../api/carto-api-error';
 import {DEFAULT_HEADERS, DEFAULT_PARAMETERS, MAX_GET_LENGTH} from './common';
-import {buildMapsUrlFromBase} from '../config';
-import type {APIErrorContext, MapType} from '../api/types';
+import type {APIErrorContext} from '../api/types';
 
 /**
  * Simple encode parameter
@@ -61,18 +60,4 @@ export async function requestWithParameters<T = any>({
   } catch (error) {
     throw new CartoAPIError(error as Error, errorContext);
   }
-}
-
-export function buildApiEndpoint({
-  apiBaseUrl,
-  connectionName,
-  endpoint,
-  mapsUrl
-}: {
-  apiBaseUrl: string;
-  connectionName: string;
-  endpoint: MapType;
-  mapsUrl?: string;
-}): string {
-  return `${mapsUrl || buildMapsUrlFromBase(apiBaseUrl)}/${connectionName}/${endpoint}`;
 }
