+++ modules/carto/src/style/color-bins-style.ts
@@ -2,7 +2,8 @@ import {scaleThreshold} from 'd3-scale';
 import {AccessorFunction} from '@deck.gl/core';
 import {Feature} from 'geojson';
 import getPalette, {Color, DEFAULT_PALETTE, NULL_COLOR} from './palette';
-import {assert, AttributeSelector, getAttrValue} from './utils';
+import {assert} from '../utils';
+import {AttributeSelector, getAttrValue} from './utils';
 
 /**
  * Helper function for quickly creating a color bins style based on `d3` `scaleThreshold`.
