+++ modules/carto/src/style/color-categories-style.ts
@@ -1,7 +1,8 @@
 import {AccessorFunction} from '@deck.gl/core';
 import {Feature} from 'geojson';
 import getPalette, {Color, DEFAULT_PALETTE, NULL_COLOR, OTHERS_COLOR} from './palette';
-import {assert, AttributeSelector, getAttrValue} from './utils';
+import {assert} from '../utils';
+import {AttributeSelector, getAttrValue} from './utils';
 
 /**
  * Helper function for quickly creating a color category style.
