+++ modules/carto/src/style/color-continuous-style.ts
@@ -2,7 +2,8 @@ import {AccessorFunction} from '@deck.gl/core';
 import {scaleLinear} from 'd3-scale';
 import {Feature} from 'geojson';
 import getPalette, {Color, DEFAULT_PALETTE, NULL_COLOR} from './palette';
-import {assert, AttributeSelector, getAttrValue} from './utils';
+import {assert} from '../utils';
+import {AttributeSelector, getAttrValue} from './utils';
 
 /**
  * Helper function for quickly creating a color continuous style.
